--- api/tests/db/base_class.py
@@ -1,3 +1,0 @@
-from sqlalchemy.orm import declarative_base
-
-Base = declarative_base()
