--- api/tests/test_api.py
@@ -1,5 +1,0 @@
-from sqlalchemy.orm import Session
-
-class TestApi:
-    """Base test class that all test classes should inherit from."""
-    pass  # Database setup is now handled by fixtures in conftest.py
